corrected summary code

+++ src/main/java/com/letswork/crm/dtos/CentreDayPassSummaryDto.java
@@ -13,6 +13,6 @@ public class CentreDayPassSummaryDto {
     
     private String state;
     
-    private Integer totalDayPassCredits;
+    private Long totalDayPassCredits;
     
 }
